move check_nix_available to start of binary
